Correccion

+++ workspace_netbeans_html/BEANS/src/net/miatech/praxis/flown/A1691.java
@@ -65,7 +65,7 @@ public class A1691 {
     public String FSENDFI = "";
     public int QCPNFI = 0;
     public int QCPNFRE = 0;
-    public int QCPHARB = 0;
+    public long QCPHARB = 0;
     public String FSTAFI = "";
     
     //tot
